+++ CHANGES.md
@@ -1,7 +1,7 @@
 # pyfakefs Release Notes
 The release versions are PyPi releases.
 
-## Version 3.6 (as yet unreleased)
+## [Version 3.5.4](https://pypi.python.org/pypi/pyfakefs/3.5.4)
 
 #### New Features
   * added context manager class `Pause` for pause/resume